Initial implementation of grant_type urn:ietf:params:oauth:grant-type:jwt-bearer
